Moved source files into single app directory
